add text into layout placeholder

add text into layout placeholder

phl_with_text(x, olay, id, str, type = "title", ...)

Arguments

x

rpptx object

olay

an OfficerLayout object created using phl_layout

id

an single integer with an id of the placeholder from olay object.

str

text to add.

type

type of the text placeholder. See ph_location_template for more details.

...

other arguments passed to ph_with.

Value

rpptx object which represents PowerPoint presentation in officer. The returned object contains a new element on the slide.