Print a CustomLayout object.

Print a CustomLayout object.

# S3 method for CustomLayout
print(x, ...)

Arguments

x
object of class CustomLayout.

...
optional arguments to print or plot methods. Not used here.

See also

Examples

lay  <- lay_new(matrix(1:4,nc=2),widths=c(3,2),heights=c(2,1))
lay2 <- lay_new(matrix(1:3))
cl <- lay_bind_col(lay,lay2, widths=c(3,1))
print(cl)

#> CustomLayout object:
#>   Specification:
#>          9    6    5
#>      1   1    3    5
#>      1   1    3    6
#>      1   2    4    7
#> 
#> 
#>   To print the layout on your graphics device plese use:
#>   lay(cl)

cl2 <- lay_bind_col(cl,cl, c(2,1))
print(cl2)

#> CustomLayout object:
#>   Specification:
#>         18   12   10    9    6    5
#>      1   1    3    5    8   10   12
#>      1   1    3    6    8   10   13
#>      1   2    4    7    9   11   14
#> 
#> 
#>   To print the layout on your graphics device plese use:
#>   lay(cl2)

cl3 <- lay_bind_row(cl,cl, c(20,1))
print(cl3)

#> CustomLayout object:
#>   Specification:
#>          9    6    5
#>     20   1    3    5
#>     20   1    3    6
#>     20   2    4    7
#>      1   8   10   12
#>      1   8   10   13
#>      1   9   11   14
#> 
#> 
#>   To print the layout on your graphics device plese use:
#>   lay(cl3)